# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Nova Mart Commerce API
  version: 1.0.0
  description: |
    Public HTTP contract for the **Nova Mart** online shopping platform.

    This document is the **source of truth**. It was authored *before* any service
    implementation, and every backend endpoint and every frontend API client method
    is derived from it. Changing behaviour means changing this file first.

    ## Architecture

    Clients never talk to a service directly. All traffic enters through the
    **API Gateway** (`/api/v1/**`), which authenticates the caller, strips any
    client-supplied identity headers, injects trusted ones, and routes to the
    owning service.

    | Path prefix                                   | Owning service       | Owns database     |
    | --------------------------------------------- | -------------------- | ----------------- |
    | `/auth`, `/users`                              | auth-service         | `auth_db`         |
    | `/products`, `/categories`, `/brands`          | product-service      | `product_db`      |
    | `/cart`                                        | cart-service         | `cart_db`         |
    | `/orders`                                      | order-service        | `order_db`        |
    | `/payments`                                    | payment-service      | `payment_db`      |
    | `/inventory`                                   | inventory-service    | `inventory_db`    |
...

...

    ## Response envelope

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: List categories
      security: []
      responses:
        '200':
          description: Categories with live product counts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Categories]
      operationId: createCategory
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Categories]
      operationId: updateCategory
      summary: Update a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Categories]
      operationId: deleteCategory
      summary: Delete a category (admin)
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Category still has products assigned to it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: Cannot delete a category that still contains products
                errorCode: CATEGORY_NOT_EMPTY
                timestamp: '2026-02-14T10:02:44.120Z'
                path: /api/v1/categories/6f1d...
  /brands:
    get:
      tags: [Brands]
      operationId: listBrands
      summary: List brands
      security: []
      responses:
        '200':
          description: Brands
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Brand'
        '500':
          $ref: '#/components/responses/ServerError'

  # =====================================================================
  #  CART
  # =====================================================================
...
  /payments:
    post:
      tags: [Payments]
      operationId: createPayment
      summary: Open a payment against an order
      description: |
        Creates an `INITIATED` payment and returns a transaction reference.
        Normally called by order-service during checkout rather than by a browser.

        **This is a simulated gateway.** No card data is accepted. `MOCK_CARD`
        settles immediately on verification; `CASH_ON_DELIVERY` stays `INITIATED`
        until the courier collects.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment opened
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Payments]
      operationId: listPayments
      summary: List payments (admin)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: Paged payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedPayments'
        '403':
          $ref: '#/components/responses/Forbidden'
  /payments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Payments]
      operationId: getPayment
      summary: Get a payment with its transaction trail
      responses:
        '200':
          description: Payment detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/{id}/verify:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Payments]
      operationId: verifyPayment
      summary: Settle a payment through the simulated gateway
      description: |
        Moves an `INITIATED` payment to `SUCCESS` or `FAILED`. The outcome is
        **deterministic, not random**, so demos and tests are reproducible:

        - `CASH_ON_DELIVERY` remains `INITIATED` and the order is still confirmed
        - a request carrying `simulateFailure: true` always yields `FAILED`
        - every other `MOCK_CARD` payment yields `SUCCESS`

        Verifying an already-settled payment is idempotent and returns the
        existing outcome rather than re-settling it.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                simulateFailure:
                  type: boolean
                  default: false
                  description: Demo hook that forces a declined payment so the failure path can be shown.
      responses:
        '200':
          description: Settlement outcome
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Payment is in a state that cannot be verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =====================================================================
  #  INVENTORY
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT issued by auth-service. Send as `Authorization: Bearer <accessToken>`.'
  parameters:
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 12
  responses:
    ValidationError:
      description: The request body or query string failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Request validation failed
            errorCode: VALIDATION_FAILED
            timestamp: '2026-02-14T09:40:12.882Z'
            path: /api/v1/auth/register
            fieldErrors:
              - field: password
                message: Password must be at least 8 characters
              - field: email
                message: Must be a well-formed email address
    Unauthorized:
      description: Missing, malformed or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication is required to access this resource
            errorCode: UNAUTHORIZED
            timestamp: '2026-02-14T09:41:03.114Z'
            path: /api/v1/cart
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: You do not have permission to perform this action
            errorCode: FORBIDDEN
            timestamp: '2026-02-14T09:42:55.900Z'
            path: /api/v1/products
    ServerError:
      description: Unexpected server-side failure. The response never contains a stack trace.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Something went wrong on our side. Please try again.
            errorCode: INTERNAL_ERROR
            timestamp: '2026-02-14T09:44:02.010Z'
            path: /api/v1/categories
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Product not found
            errorCode: PRODUCT_NOT_FOUND
            timestamp: '2026-02-14T09:43:30.720Z'
            path: /api/v1/products/8f14e45f-ea8f-4b2a-9d1c-3c2b7a51de90
  schemas:
...
    ApiResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Product retrieved successfully
        data:
          description: Operation-specific payload. Absent for `204` responses.
    ErrorResponse:
      type: object
      required: [success, message, errorCode, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
          description: Human-readable, safe to show a shopper. Never contains a stack trace.
        errorCode:
          type: string
          description: Stable machine-readable identifier. Clients branch on this, not on the message text.
          example: PRODUCT_NOT_FOUND
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        fieldErrors:
          type: array
          description: Present only when `errorCode` is `VALIDATION_FAILED`.
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, first, last]
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 12
        totalElements:
          type: integer
          format: int64
          example: 24
        totalPages:
          type: integer
          example: 2
        first:
          type: boolean
        last:
          type: boolean

    # ---------- auth ----------
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          format: uri
          nullable: true
        productCount:
          type: integer
    CategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          maxLength: 500
          nullable: true
        imageUrl:
          type: string
          format: uri
          nullable: true
    Brand:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        productCount:
          type: integer
...
    PaymentMethod:
      type: string
      enum: [MOCK_CARD, CASH_ON_DELIVERY]
      description: |
        `MOCK_CARD` is a simulated gateway - no card details are collected or stored.
        `CASH_ON_DELIVERY` defers capture until handover.
    PaymentStatus:
      type: string
      enum: [INITIATED, SUCCESS, FAILED, REFUNDED]
    CreatePaymentRequest:
      type: object
      required: [orderId, amount, currency, method]
      properties:
        orderId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        currency:
          type: string
          example: INR
        method:
          $ref: '#/components/schemas/PaymentMethod'
    Payment:
      type: object
      required: [id, orderId, amount, currency, method, status, transactionReference, createdAt]
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        currency:
          type: string
        method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        transactionReference:
          type: string
          example: NMPAY-7QK3M2XB
        failureReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        settledAt:
          type: string
          format: date-time
          nullable: true
    PaymentDetail:
      allOf:
        - $ref: '#/components/schemas/Payment'
        - type: object
          properties:
            transactions:
              type: array
              items:
                $ref: '#/components/schemas/Transaction'
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [AUTHORIZE, CAPTURE, REFUND, DECLINE]
        amount:
          type: number
          format: double
        gatewayReference:
          type: string
        message:
          type: string
        occurredAt:
          type: string
          format: date-time
    PagedPayments:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        page:
          $ref: '#/components/schemas/PageMeta'

    # ---------- inventory ----------
...